updates validator test class
